Compatibility for macos

+++ client/src/workers/usingWorker.ts
@@ -2,13 +2,19 @@ import memoize from "memoizee";
 
 type WorkerConstructor = new () => Worker;
 
+type WorkerResult = { result: any } | { error: any };
+
 export const usingWorker =
   <R>(w: WorkerConstructor) =>
   async (task: (w: Worker) => Promise<R>) => {
     const worker = new w();
-    const out = await task(worker);
+    const out = (await task(worker)) as WorkerResult;
+    if ("error" in out) {
+      console.error(out.error);
+      throw new Error(out.error);
+    }
     worker.terminate();
-    return out;
+    return out.result as R;
   };
 
 export const usingWorkerTask =
@@ -20,7 +26,10 @@ export const usingWorkerTask =
         worker.onmessage = (out) => {
           res(out.data as R);
         };
-        worker.onerror = rej;
+        worker.onerror = (e) => {
+          console.error(e);
+          rej(e);
+        };
       });
     });
 
@@ -31,3 +40,14 @@ export const usingMemoizedWorkerTask = <T, R>(
     length: 1,
   }
 ) => memoize(usingWorkerTask(w), o);
+
+export const usingMessageHandler =
+  <T, U>(f: (a: MessageEvent<T>) => Promise<U>) =>
+  async (m: MessageEvent<T>) => {
+    try {
+      const output = await f(m);
+      postMessage({ result: output });
+    } catch (e) {
+      postMessage({ error: e });
+    }
+  };
